server/requirements.txt
channels-redis==2.4.1
channels==2.4.0
django-concurrency==2.2
django-cors-headers==3.2.1
django-sequences==2.4
django-templated-mail==1.1.1
Django==3.0.10
djangorestframework==3.11.1
djoser==2.1.0